Moved device struct for batched gemm wmma to a common file
